Add helpers for deprecating modules.
This introduces two new functions for deprecating modules:
`module_deprecated()` and `module_moved()`.

`module_deprecated()` can be called in the body of a module to emit a
deprecation warning when the module is imported.

`module_moved()` is similar, but takes a destination module, and points
consumers to that module in the deprecation message.

Testing Done:
Unit tests pass on all supported versions of Python.

Reviewed at https://reviews.reviewboard.org/r/13014/