Add version info, mark as supporting typing, and use setuptools.
We now store Housekeeping version information in
`housekeeping._version`. This is now used when building the package,
instead of hard-coding a version. For that, we're now using the
setuptools build backend for `pyproject.toml`.

We also store a `py.typed` file in the module, which indicates to
consumers that this package provides type hints.

Testing Done:
Successfully built packages. Verified that consumers could see typing
information.

Reviewed at https://reviews.reviewboard.org/r/13030/